requirements.txt
affine==2.4.0
attrs==23.1.0
certifi==2023.5.7
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
Fiona==1.9.4.post1
geopandas==0.13.0
importlib-metadata==6.6.0
numpy==1.24.3
packaging==23.1
pandas==2.0.1
pyparsing==3.0.9
pyproj==3.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
rasterio==1.3.7
shapely==2.0.1
six==1.16.0
snuggs==1.4.7
tzdata==2023.3
zipp==3.15.0
